Fix previous output comparison for build stats formatter
