+++ .gitignore
@@ -30,3 +30,6 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# tmp files
+.tmp
